backend/requirements.txt
asgiref==3.4.1
attrs==21.2.0
black==21.10b0
click==8.0.3
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
flake8==4.0.1
freezegun==1.1.0
iniconfig==1.1.1
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
packaging==21.2
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2021.3
regex==2021.11.10
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
